backend/requirements.txt
asgiref==3.11.0
Django==5.2.8
django-tinymce==5.0.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
sqlparse==0.5.3
typing_extensions==4.15.0
boto3==1.41.2
botocore==-1.41.2 
requests==2.32.5
